+++ macos/Classes/AutoUpdaterPlugin.swift
@@ -42,7 +42,7 @@ public class AutoUpdaterPlugin: NSObject, FlutterPlugin {
             result(true)
             break
         case "setScheduledCheckInterval":
-            let interval = args["interval"] as! Double
+            let interval = args["interval"] as! Int
             autoUpdater.setScheduledCheckInterval(interval)
             result(true)
             break
